# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/visits": {
            "get": {
                "tags": [
                    "Visits"
                ],
                "summary": "Get list of visits",
                "operationId": "59bc502e31c31bcb9d32d86e06260de5",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Visit"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Visits"
                ],
                "summary": "Create a new visit",
                "operationId": "2ff60d5091a706b6942379521ae9074f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Visit"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Visit created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Visit"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/visits/{id}": {
            "get": {
                "tags": [
                    "Visits"
                ],
                "summary": "Get visit details",
                "operationId": "42d6ce6dce87180d7400f1a2ad20259e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Visit ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Visit"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Visit not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Visits"
                ],
                "summary": "Update a visit",
                "operationId": "460d80e475b87e62bf034f944d29cacf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Visit ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Visit"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Visit updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Visit"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Visits"
                ],
                "summary": "Delete a visit",
                "operationId": "17961e50292ffea006c5ff414d95307b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Visit ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Visit deleted successfully"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Visit": {
                "properties": {
                    "id": {
                        "description": "The visit ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the visit",
                        "type": "string"
                    },
                    "email": {
                        "description": "The email of the visit",
                        "type": "string"
                    },
                    "latitude": {
                        "description": "The latitude of the visit",
                        "type": "string"
                    },
                    "longitude": {
                        "description": "The longitude of the visit",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "The creation timestamp of the visit",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The update timestamp of the visit",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Visits",
            "description": "API Endpoints of Visits"
        },
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}